gateway: remove thread handling
