+++ cmd/buckd/docker-compose-dev.yml
@@ -16,6 +16,7 @@ services:
       - BUCK_GATEWAY_SUBDOMAINS
       - BUCK_GATEWAY_WWW_DOMAIN
       - BUCK_THREADS_ADDR=threads:5000
+      - BUCK_THREADS_GATEWAY_URL=http://127.0.0.1:7000
       - BUCK_IPFS_MULTIADDR=/dns4/ipfs/tcp/5001
       - BUCK_IPNS_REPUBLISH_SCHEDULE
       - BUCK_IPNS_REPUBLISH_CONCURRENCY
@@ -34,7 +35,7 @@ services:
       - threads
       - ipfs
   threads:
-    image: textile/go-threads:534a6d0
+    image: textile/go-threads:1ef50f7
     restart: always
     volumes:
       - "${REPO_PATH}/threads:/data/threads"
@@ -42,12 +43,14 @@ services:
       - THREADS_DEBUG=true
       - THREADS_APIADDR=/ip4/0.0.0.0/tcp/5000
       - THREADS_APIPROXYADDR=/ip4/0.0.0.0/tcp/5050
+      - THREADS_GATEWAYADDR=0.0.0.0:8000
     ports:
       - "4066:4006"
       - "4066:4006/udp"
       - "127.0.0.1:4050:5050"
+      - "127.0.0.1:7000:8000"
   ipfs:
-    image: ipfs/go-ipfs:v0.8.0
+    image: textile/go-ipfs:v0.8.0-m1
     volumes:
       - "${REPO_PATH}/ipfs:/data/ipfs"
     environment:
